+++ build.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-
 ulimit -n 100000;
 composer update --no-dev;
 php ./phing.phar release;